fix(ProgressBar): corrected typescript definition HVUIKIT-5670
